Division by ContextFullCommand and ContextLess Commands
