+++ tg-bot/src/main/java/polis/commands/ContextLessCommand.java
@@ -1,0 +1,6 @@
+package polis.commands;
+
+import org.telegram.telegrambots.extensions.bots.commandbot.commands.IBotCommand;
+
+public interface ContextLessCommand extends IBotCommand {
+}
