+++ tg-bot/src/main/java/polis/commands/contextfull/AddOkAccount.java
@@ -1,4 +1,4 @@
-package polis.commands;
+package polis.commands.contextfull;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -6,6 +6,7 @@ import org.springframework.stereotype.Component;
 import org.telegram.telegrambots.meta.api.objects.Chat;
 import org.telegram.telegrambots.meta.api.objects.User;
 import org.telegram.telegrambots.meta.bots.AbsSender;
+import polis.commands.Command;
 import polis.keyboards.InlineKeyboard;
 import polis.keyboards.ReplyKeyboard;
 import polis.ok.api.OkAuthorizator;
