+++ tg-bot/src/main/java/polis/commands/contextfull/AddVkGroup.java
@@ -1,10 +1,11 @@
-package polis.commands;
+package polis.commands.contextfull;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.telegram.telegrambots.meta.api.objects.Chat;
 import org.telegram.telegrambots.meta.api.objects.User;
 import org.telegram.telegrambots.meta.bots.AbsSender;
+import polis.commands.Command;
 import polis.data.domain.ChannelGroup;
 import polis.data.domain.CurrentChannel;
 import polis.data.repositories.ChannelGroupsRepository;
@@ -15,12 +16,12 @@ import polis.util.SocialMedia;
 import polis.util.State;
 
 @Component
-public class AddOkGroup extends Command {
-    private static final String ADD_OK_GROUP_MSG = """
+public class AddVkGroup extends Command {
+    private static final String ADD_VK_GROUP_MSG = """
             Чтобы добавить новую группу, введите в чат ссылку на нее.
             Примеры ссылок:
-            https://ok.ru/ok
-            https://ok.ru/group57212027273260""";
+            https://vk.com/lentach
+            https://vk.com/club1234567890""";
     static final String SAME_SOCIAL_MEDIA_MSG = """
             Социальная сеть %s уже была синхронизирована с текущим Телеграмм-каналом.
             Пожалуйста, выберите другую социальную сеть и попробуйте снова.""";
@@ -31,8 +32,8 @@ public class AddOkGroup extends Command {
     @Autowired
     private ChannelGroupsRepository channelGroupsRepository;
 
-    public AddOkGroup(InlineKeyboard inlineKeyboard, ReplyKeyboard replyKeyboard) {
-        super(State.AddOkGroup.getIdentifier(), State.AddOkGroup.getDescription(), inlineKeyboard, replyKeyboard);
+    public AddVkGroup(InlineKeyboard inlineKeyboard, ReplyKeyboard replyKeyboard) {
+        super(State.AddVkGroup.getIdentifier(), State.AddVkGroup.getDescription(), inlineKeyboard, replyKeyboard);
     }
 
     @Override
@@ -40,11 +41,11 @@ public class AddOkGroup extends Command {
         CurrentChannel currentChannel = currentChannelRepository.getCurrentChannel(chat.getId());
         if (currentChannel != null) {
             for (ChannelGroup smg : channelGroupsRepository.getGroupsForChannel(currentChannel.getChannelId())) {
-                if (smg.getSocialMedia() == SocialMedia.OK) {
+                if (smg.getSocialMedia() == SocialMedia.VK) {
                     sendAnswerWithOnlyBackButton(
                             absSender,
                             chat.getId(),
-                            String.format(SAME_SOCIAL_MEDIA_MSG, SocialMedia.OK.getName()),
+                            String.format(SAME_SOCIAL_MEDIA_MSG, SocialMedia.VK.getName()),
                             loggingInfo(user.getUserName()));
                     return;
                 }
@@ -53,7 +54,7 @@ public class AddOkGroup extends Command {
         sendAnswerWithOnlyBackButton(
                 absSender,
                 chat.getId(),
-                ADD_OK_GROUP_MSG,
+                ADD_VK_GROUP_MSG,
                 loggingInfo(user.getUserName()));
     }
 }
