+++ tg-bot/src/main/java/polis/commands/contextfull/OkAccountDescription.java
@@ -1,10 +1,11 @@
-package polis.commands;
+package polis.commands.contextfull;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.telegram.telegrambots.meta.api.objects.Chat;
 import org.telegram.telegrambots.meta.api.objects.User;
 import org.telegram.telegrambots.meta.bots.AbsSender;
+import polis.commands.Command;
 import polis.data.domain.CurrentAccount;
 import polis.data.repositories.CurrentAccountRepository;
 import polis.keyboards.InlineKeyboard;
