+++ tg-bot/src/main/java/polis/commands/contextless/Help.java
@@ -1,15 +1,17 @@
-package polis.commands;
+package polis.commands.contextless;
 
 import org.springframework.stereotype.Component;
 import org.telegram.telegrambots.meta.api.objects.Chat;
 import org.telegram.telegrambots.meta.api.objects.User;
 import org.telegram.telegrambots.meta.bots.AbsSender;
+import polis.commands.Command;
+import polis.commands.ContextLessCommand;
 import polis.keyboards.InlineKeyboard;
 import polis.keyboards.ReplyKeyboard;
 import polis.util.State;
 
 @Component
-public class Help extends Command {
+public class Help extends Command implements ContextLessCommand {
     private static final String HELP = """
             <b>Feed-Migrator</b> - бот авто-постинга в Телеграмме.
             Бот позволяет подключить несколько аккаунтов с группами в социальных сетях ВКонтакте и Одноклассники к"""
